commited all previous movements
